Implement user's profile upload, refresh token not found problem et al.

+++ src/main/java/com/ous/aethererp/service/RefreshTokenServiceImpl.java
@@ -73,7 +73,11 @@ public class RefreshTokenServiceImpl implements RefreshTokenService {
     public RefreshTokenEntity verifyRefreshToken(String token) {
         RefreshTokenEntity refreshToken=
                 refreshTokenRepo.findByToken(token)
-                        .orElseThrow(() -> new RuntimeException("Refresh token not found."));
+                        .orElseThrow(() -> new RuntimeException("Refresh token not found"));
+
+        if (refreshToken.isRevoked()){
+            throw new RuntimeException("Refresh token has been revoked.");
+        }
 
         if (refreshToken.getExpiryDate() < System.currentTimeMillis()){
             refreshTokenRepo.delete(refreshToken);
